Move AppWidgetHelper to library
