+++ features/app-widgets/build.gradle.kts
@@ -47,6 +47,7 @@ dependencies {
     implementation(projects.common.designCompose)
     implementation(projects.common.lessonsData)
     implementation(projects.common.translations)
+    implementation(projects.libraries.appWidget.api)
     implementation(projects.libraries.foundation.coroutines)
     implementation(projects.libraries.lessons.api)
     implementation(projects.libraries.prefs.api)
