+++ features/app-widgets/src/main/java/app/ss/widgets/di/WidgetModule.kt
@@ -1,6 +1,5 @@
 package app.ss.widgets.di
 
-import app.ss.widgets.AppWidgetHelper
 import app.ss.widgets.AppWidgetHelperImpl
 import app.ss.widgets.data.AppWidgetRepository
 import app.ss.widgets.data.AppWidgetRepositoryImpl
@@ -8,6 +7,7 @@ import dagger.Binds
 import dagger.Module
 import dagger.hilt.InstallIn
 import dagger.hilt.components.SingletonComponent
+import ss.libraries.appwidget.api.AppWidgetHelper
 
 @Module
 @InstallIn(SingletonComponent::class)
