+++ features/lessons/build.gradle.kts
@@ -20,32 +20,6 @@
  * THE SOFTWARE.
  */
 
-import com.android.build.gradle.internal.tasks.databinding.DataBindingGenBaseClassesTask
-import org.gradle.configurationcache.extensions.capitalized
-import org.jetbrains.kotlin.gradle.tasks.AbstractKotlinCompileTool
-
-/*
- * Copyright (c) 2023. Adventech <info@adventech.io>
- *
- * Permission is hereby granted, free of charge, to any person obtaining a copy
- * of this software and associated documentation files (the "Software"), to deal
- * in the Software without restriction, including without limitation the rights
- * to use, copy, modify, merge, publish, distribute, sublicense, and/or sell
- * copies of the Software, and to permit persons to whom the Software is
- * furnished to do so, subject to the following conditions:
- *
- * The above copyright notice and this permission notice shall be included in
- * all copies or substantial portions of the Software.
- *
- * THE SOFTWARE IS PROVIDED "AS IS", WITHOUT WARRANTY OF ANY KIND, EXPRESS OR
- * IMPLIED, INCLUDING BUT NOT LIMITED TO THE WARRANTIES OF MERCHANTABILITY,
- * FITNESS FOR A PARTICULAR PURPOSE AND NON-INFRINGEMENT. IN NO EVENT SHALL THE
- * AUTHORS OR COPYRIGHT HOLDERS BE LIABLE FOR ANY CLAIM, DAMAGES OR OTHER
- * LIABILITY, WHETHER IN AN ACTION OF CONTRACT, TORT OR OTHERWISE, ARISING FROM,
- * OUT OF OR IN CONNECTION WITH THE SOFTWARE OR THE USE OR OTHER DEALINGS IN
- * THE SOFTWARE.
- */
-
 plugins {
     alias(libs.plugins.sgp.base)
     alias(libs.plugins.android.library)
@@ -69,8 +43,8 @@ dependencies {
     implementation(projects.common.designCompose)
     implementation(projects.common.translations)
     implementation(projects.common.lessonsData)
-    implementation(projects.features.appWidgets)
     implementation(projects.features.pdf)
+    implementation(projects.libraries.appWidget.api)
     implementation(projects.libraries.circuit.api)
     implementation(projects.libraries.foundation.android)
     implementation(projects.libraries.foundation.coroutines)
@@ -95,6 +69,7 @@ dependencies {
     implementation(libs.markwon.core)
 
     testImplementation(libs.bundles.testing.common)
+    testImplementation(projects.libraries.appWidget.testFixtures)
     testImplementation(projects.libraries.testUtils)
     testImplementation(projects.libraries.foundation.coroutines.test)
     testImplementation(projects.libraries.lessons.test)
