+++ libraries/app-widget/api/build.gradle.kts
@@ -1,0 +1,5 @@
+plugins {
+    alias(libs.plugins.sgp.base)
+    alias(libs.plugins.android.library)
+    alias(libs.plugins.kotlin.android)
+}
