+++ libraries/app-widget/api/src/main/kotlin/ss/libraries/appwidget/api/AppWidgetHelper.kt
@@ -1,0 +1,7 @@
+package ss.libraries.appwidget.api
+
+interface AppWidgetHelper {
+    fun refreshAll()
+    suspend fun isAdded(): Boolean
+    fun syncQuarterly(index: String)
+}
