+++ libraries/app-widget/test-fixtures/build.gradle.kts
@@ -1,0 +1,9 @@
+plugins {
+    alias(libs.plugins.sgp.base)
+    alias(libs.plugins.android.library)
+    alias(libs.plugins.kotlin.android)
+}
+
+dependencies {
+    implementation(projects.libraries.appWidget.api)
+}
